% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_rlang.R
\name{append_rlang}
\alias{append_rlang}
\alias{append_rlang.function_usage}
\alias{append_rlang.package_usage}
\title{Append rlang !!sym}
\usage{
append_rlang(obj)

\method{append_rlang}{function_usage}(obj)

\method{append_rlang}{package_usage}(obj)
}
\arguments{
\item{obj}{function_usage or package_usage class}
}
\value{
function
}
\description{
Programatically append !!rlang::sym('[OBJECT]') to the body of a function
}
\examples{
\dontrun{
x <- function(){
  
  data <- tidyr::unite(mtcars, col = vs_am, c(vs,am))
  
  ggplot2::ggplot(data = data, ggplot2::aes(x=mpg^2,y=qsec,colour=vs_am)) + 
    ggplot2::geom_point()
}

obj <- tidycheckUsage::tidycheckUsage(x)

(x1 <- tidycheckUsage::append_rlang(obj))

tidycheckUsage::tidycheckUsage(x1)

x1()
}
}
