% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_rlang_package.R
\name{append_rlang_package}
\alias{append_rlang_package}
\title{Append rlang !!sym on package}
\usage{
append_rlang_package(pack, unquo_type = c("UQ", "!!"), ...)
}
\arguments{
\item{pack}{character, name of package to check.}

\item{unquo_type}{character, unquo type, Default: 'UQ'}

\item{...}{options to be passed to checkUsagePackage}
}
\description{
Programatically append !!rlang::sym('[OBJECT]') to the body of a function
}
\details{
By default rlang::UQ is used because there are still
 problems with '!!', eg !!rlang::sym(x)^2 will cause an error.
}
