% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUsage_dataframe.R
\name{checkUsage_dataframe}
\alias{checkUsage_dataframe}
\title{return checkUsage as data.frame}
\usage{
checkUsage_dataframe(fun, ...)
}
\arguments{
\item{fun}{closure, function to check}

\item{...}{options to be passed to checkUsage}
}
\value{
data.frame
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
myfun <- function(x){

ret <- mtcars\%>\%
  mutate(mpg2=mpg*x)

ret <- ret\%>\%
  mutate(mpg3=mpg2^2)

}

checkUsage_dataframe(fun = myfun)

}
\seealso{
\code{\link[codetools]{checkUsage}}
}
