% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycheckUsagePackage.R
\name{tidycheckUsagePackage}
\alias{tidycheckUsagePackage}
\title{FUNCTION_TITLE}
\usage{
tidycheckUsagePackage(pack, ...)
}
\arguments{
\item{pack}{character, name of package to check.}

\item{...}{options to be passed to checkUsage.}
}
\value{
data.frame
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
\code{\link[codetools]{checkUsage}}
}
