% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null-default}
\alias{null-default}
\alias{\%||\%}
\title{Default NULL}
\usage{
x \%||\% y
}
\arguments{
\item{x, y}{If `x` is NULL, will return `y`; otherwise returns `x`.}
}
\description{
Infix function to replace NULLs
}
