% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Display checkUsage results using a standalone report}
\usage{
report(x = tidycheckUsagePackage(), file = file.path(tempdir(),
  paste0(get_pack_name(x), "-report.html")), browse = interactive())
}
\arguments{
\item{x}{a tidycheckUsage dataset, defaults to running `checkUsagePackage_dataframe()`.}

\item{file}{The report filename.}

\item{browse}{whether to open a browser to view the report.}
}
\description{
Display checkUsage results using a standalone report
}
\examples{
\dontrun{
x <- tidycheckUsagePackage()
report(x)
}
}
