% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycheckUsagePackage.R
\name{tidycheckUsagePackage}
\alias{tidycheckUsagePackage}
\title{run tidycheckUsage on an installed package}
\usage{
tidycheckUsagePackage(pack, markers = TRUE, ...)
}
\arguments{
\item{pack}{character, name of package to check.}

\item{markers}{boolean, invoke RStudio markers, Default: TRUE}

\item{...}{options to be passed to checkUsage.}
}
\value{
data.frame
}
\description{
Evaluates tidycheckUsage on entire package
}
\details{
by default the skipWith arguments of checkUsagePackage
is set to TRUE to mimic the devtools::check() function.
}
\seealso{
\code{\link[codetools]{checkUsage}}
}
