% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_rlang.R
\name{append_rlang}
\alias{append_rlang}
\title{Append rlang !!sym}
\usage{
append_rlang(.f, unquo_type = c("UQ", "!!"), ...)
}
\arguments{
\item{.f}{function}

\item{unquo_type}{character, unquo type, Default: 'UQ'}

\item{...}{options to be passed to checkUsage.}
}
\value{
function
}
\description{
Programatically append !!rlang::sym('[OBJECT]') to the body of a function
}
\examples{
\dontrun{
x <- function(){
  
  data <- tidyr::unite(mtcars, col = vs_am, c(vs,am))
  
  ggplot2::ggplot(data = data, ggplot2::aes(x=mpg^2,y=qsec,colour=vs_am)) + 
    ggplot2::geom_point()
}

tidycheckUsage::tidycheckUsage(x)

(x1 <- tidycheckUsage::append_rlang(x))

tidycheckUsage::tidycheckUsage(x1)

x1()
}
}
