% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUsagePackage_dataframe.R
\name{checkUsagePackage_dataframe}
\alias{checkUsagePackage_dataframe}
\title{FUNCTION_TITLE}
\usage{
checkUsagePackage_dataframe(pack, ...)
}
\arguments{
\item{pack}{character, name of package to check.}

\item{...}{options to be passed to checkUsage.}
}
\value{
data.frame
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
\code{\link[codetools]{checkUsage}}
}
