% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silent_check.R
\name{silent_check}
\alias{silent_check}
\title{Run devtools::check in a backaground session}
\usage{
silent_check(...)
}
\arguments{
\item{\dots}{Arguments to pass to \code{\link[devtools]{check}}}
}
\value{
An object containing errors, warnings, and notes.
}
\description{
Run \code{\link[devtools]{check}} in another R session
}
\details{
Silent option of devtools::check is foreced to TRUE.
}
\seealso{
\code{\link[callr]{r_bg}}
 \code{\link[devtools]{check}}
}
